Complete re-write of the list_dobatch process for both mdtest and ior
The testing is now driven out of four files: -f (filesystems) -p
(processes - a list of the number of processes for each run) -o (options
to specify any run-time options to the respective benchmark -r (runner)
the options passed to the md_runner or ior_runner script to control the
benchmark environment.

There are two new scripts affiliated with the md_runner script that will
clean up any directory trees that may have been orphaned by a prior run.
This cleanup process will run in the background and will not have an
impact on other benchmarks running at the same time.